requirements.txt
numpy==1.26.4
pillow==10.4.0
tiffslide==2.4.0
tiffslide_xarray==0.2.1
scikit-image==0.24.0
scikit-learn==1.5.1
matplotlib==3.9.1.post1
pandas==1.5.3
tqdm==4.63.0
patchify==0.2.3
xarray==2024.3.0
scipy==1.14.0
torchstain==1.3.0
seaborn==0.11.2
statannotations==0.6.0